requirements.txt
tensorboardX~=2.6.2.2
# torch                     2.1.0+cu118
# torchaudio                2.1.0+cu118
# torchvision               0.16.0+cu118

torch==2.1.0
numpy~=1.24.4
matplotlib~=3.7.3
torchvision==0.16.0
Pillow==10.1.0
PyQt5~=5.15.10 # UI界面
netron~=7.2.9 # 用于可视化onnx模型结构

opencv-contrib-python~=4.8.1.78

scikit-learn~=1.3.2
pandas~=2.0.3